{"schema":"decodex/commit/1","summary":"Modularize manual config and tracker boundaries","authority":"manual"}
